#9 move digit parsing / guessing functionality
