+++ BankOcr/BankOcr.Business.Tests/Models/OcrDigit/ConvertOcrDigitToNumber.cs
@@ -1,14 +1,14 @@
-﻿namespace BankOcr.Business.Tests.Ocr;
+﻿namespace BankOcr.Business.Tests.Models.OcrDigit;
 
 [TestFixture]
-public class ConvertOcrDigitToNumber: OcrServiceBase
+public class ConvertOcrDigitToNumber : OcrDigitBase
 {
     [TestCaseSource(nameof(GetOcrDigitTestCaseData))]
     public void Should_Return_ExpectedDigit(string ocrString, char expectedDigit)
     {
         // Act
-        var result = GetService().ConvertOcrDigitToNumber(ocrString);
-        
+        var result = Business.Models.OcrDigit.ConvertOcrDigitToNumber(ocrString);
+
         // Assert
         Assert.That(result, Is.EqualTo(expectedDigit));
     }
@@ -17,7 +17,7 @@ public class ConvertOcrDigitToNumber: OcrServiceBase
     public void Should_Return_IllegibleChar_WhenDigitIsIllegible(string ocrString)
     {
         // Act
-        var result = GetService().ConvertOcrDigitToNumber(ocrString);
+        var result = Business.Models.OcrDigit.ConvertOcrDigitToNumber(ocrString);
 
         // Assert
         Assert.That(result, Is.EqualTo('?'));
