+++ BankOcr/BankOcr.Business.Tests/Services/Ocr/OcrServiceBase.cs
@@ -4,7 +4,7 @@ using BankOcr.Business.Models;
 using BankOcr.Business.Services;
 using NSubstitute;
 
-namespace BankOcr.Business.Tests.Ocr;
+namespace BankOcr.Business.Tests.Services.Ocr;
 
 public class OcrServiceBase
 {
@@ -24,90 +24,6 @@ public class OcrServiceBase
     protected void SetAccountNumberServiceGetValidAccountNumbers(List<string> accountNumbers)
     {
         MockAccountNumberService.GetValidAccountNumbers(Arg.Any<List<string>>()).Returns(accountNumbers);
-    }
-
-    public static IEnumerable GetOcrDigitTestCaseData()
-    {
-        yield return new TestCaseData
-        (
-            " _ \n" +
-            "| |\n" +
-            "|_|\n" +
-            "",
-            '0'
-        ).SetName("0");
-        yield return new TestCaseData
-        (
-            "   \n" +
-            "  |\n" +
-            "  |\n" +
-            "",
-            '1'
-        ).SetName("1");
-        yield return new TestCaseData
-        (
-            " _ \n" +
-            " _|\n" +
-            "|_ \n" +
-            "",
-            '2'
-        ).SetName("2");
-        yield return new TestCaseData
-        (
-            " _ \n" +
-            " _|\n" +
-            " _|\n" +
-            "",
-            '3'
-        ).SetName("3");
-        yield return new TestCaseData
-        (
-            "   \n" +
-            "|_|\n" +
-            "  |\n" +
-            "",
-            '4'
-        ).SetName("4");
-        yield return new TestCaseData
-        (
-            " _ \n" +
-            "|_ \n" +
-            " _|\n" +
-            "",
-            '5'
-        ).SetName("5");
-        yield return new TestCaseData
-        (
-            " _ \n" +
-            "|_ \n" +
-            "|_|\n" +
-            "",
-            '6'
-        ).SetName("6");
-        yield return new TestCaseData
-        (
-            " _ \n" +
-            "  |\n" +
-            "  |\n" +
-            "",
-            '7'
-        ).SetName("7");
-        yield return new TestCaseData
-        (
-            " _ \n" +
-            "|_|\n" +
-            "|_|\n" +
-            "",
-            '8'
-        ).SetName("8");
-        yield return new TestCaseData
-        (
-            " _ \n" +
-            "|_|\n" +
-            " _|\n" +
-            "",
-            '9'
-        ).SetName("9");
     }
 
     public static IEnumerable GetIllegibleOcrDigitTestCaseData()
@@ -234,7 +150,7 @@ public class OcrServiceBase
             "  |  |  |  |  |  |  |  |  |\n" +
             "  |  |  |  |  |  |  |  |  |\n" +
             "\n",
-            "711111111", 
+            "711111111",
             true
         ).SetName("Valid 711111111");
         // 123456789
@@ -243,7 +159,7 @@ public class OcrServiceBase
             "  | _| _||_||_ |_   ||_||_|\n" +
             "  ||_  _|  | _||_|  ||_| _|\n" +
             "\n",
-            "123456789", 
+            "123456789",
             true
         ).SetName("Valid 123456789");
         // 490867715
@@ -252,7 +168,7 @@ public class OcrServiceBase
             "|_||_|| ||_||_   |  |  ||_ \n" +
             "  | _||_||_||_|  |  |  | _|\n" +
             "\n",
-            "490867715", 
+            "490867715",
             true
         ).SetName("Valid 490867715");
 
@@ -263,7 +179,7 @@ public class OcrServiceBase
             "|_||_||_||_||_||_||_||_||_|\n" +
             "|_||_||_||_||_||_||_||_||_|\n" +
             "\n",
-            "888888888", 
+            "888888888",
             false
         ).SetName("Invalid 888888888");
         // 490067715
@@ -272,7 +188,7 @@ public class OcrServiceBase
             "|_||_|| || ||_   |  |  ||_ \n" +
             "  | _||_||_||_|  |  |  | _|\n" +
             "\n",
-            "490067715", 
+            "490067715",
             false
         ).SetName("Invalid 490067715");
         // 012345678
@@ -281,94 +197,9 @@ public class OcrServiceBase
             "| |  | _| _||_||_ |_   ||_|\n" +
             "|_|  ||_  _|  | _||_|  ||_|\n" +
             "\n",
-            "012345678", 
+            "012345678",
             false
         ).SetName("Invalid 012345678");
-    }
-
-    public static IEnumerable GetOcrDigitInvalidGuessTestCaseData()
-    {
-        // digit that are valid but could be other numbers 
-        yield return new TestCaseData(
-            " _ \n" +
-            "| |\n" +
-            "|_|\n" +
-            "",
-            new List<char> { '8' }
-        ).SetName("0 could be 8");
-
-        yield return new TestCaseData(
-            "   \n" +
-            "  |\n" +
-            "  |\n" +
-            "",
-            new List<char> { '7' }
-        ).SetName("1 could be 7");
-
-        yield return new TestCaseData(
-            " _ \n" +
-            " _|\n" +
-            " _|\n" +
-            "",
-            new List<char> { '9' }
-        ).SetName("3 could be 9");
-
-        yield return new TestCaseData(
-            " _ \n" +
-            "|_ \n" +
-            "|_|\n" +
-            "",
-            new List<char> { '5', '8' }
-        ).SetName("6 could be 5 or 8");
-
-        yield return new TestCaseData(          
-            " _ \n" +
-            "  |\n" +
-            "  |\n" +
-            "",
-            new List<char> {'1'}
-        ).SetName("7 could be 1");
-
-        yield return new TestCaseData(
-            " _ \n" +
-            "|_|\n" +
-            "|_|\n" +
-            "",
-            new List<char> { '0', '6', '9' }
-        ).SetName("8 could be 0, 6 or 9");
-
-        yield return new TestCaseData(
-            " _ \n" +
-            "|_|\n" +
-            " _|\n" +
-            "",
-            new List<char> { '3', '5', '8' }
-        ).SetName("9 could be 3, 5 or 8");
-
-        // digits that are invalid
-        yield return new TestCaseData(
-            "   \n" +
-            " _|\n" +
-            "  |\n" +
-            "",
-            new List<char> { '1', '4' }
-        ).SetName("Invalid #1 could be 1 or 4");
-
-        yield return new TestCaseData(
-            "   \n" +
-            "| |\n" +
-            "|_|\n" +
-            "",
-            new List<char> { '0' }
-        ).SetName("Invalid #2 could be 0");
-
-        yield return new TestCaseData(
-            " _ \n" +
-            " _ \n" +
-            " _|\n" +
-            "",
-            new List<char> { '3', '5' }
-        ).SetName("Invalid #3 could be 3 or 5");
     }
 
     public static IEnumerable GetOcrNumberTestCaseData()
@@ -380,7 +211,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "000000000",
                     Status = AccountNumberStatus.Ok
@@ -396,7 +227,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "711111111",
                     Status = AccountNumberStatus.Ok
@@ -412,7 +243,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "777777177",
                     Status = AccountNumberStatus.Ok
@@ -428,7 +259,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "200800000",
                     Status = AccountNumberStatus.Ok
@@ -444,7 +275,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "333393333",
                     Status = AccountNumberStatus.Ok
@@ -460,7 +291,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "888888888",
                     Status = AccountNumberStatus.Ambiguous
@@ -477,7 +308,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "555555555",
                     Status = AccountNumberStatus.Ambiguous
@@ -494,7 +325,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "666666666",
                     Status = AccountNumberStatus.Ambiguous
@@ -511,7 +342,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "999999999",
                     Status = AccountNumberStatus.Ambiguous
@@ -528,7 +359,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "490067715",
                     Status = AccountNumberStatus.Ambiguous
@@ -545,7 +376,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "123456789",
                     Status = AccountNumberStatus.Ok
@@ -561,7 +392,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "000000051",
                     Status = AccountNumberStatus.Ok
@@ -577,7 +408,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "490867715",
                     Status = AccountNumberStatus.Ok
@@ -593,7 +424,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "000000031",
                     Status = AccountNumberStatus.Error
@@ -609,7 +440,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "810000031",
                     Status = AccountNumberStatus.Error
@@ -625,7 +456,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "970000051",
                     Status = AccountNumberStatus.Error
@@ -642,7 +473,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "0000000?1",
                     Status = AccountNumberStatus.Illegible
@@ -658,7 +489,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "?000000?1",
                     Status = AccountNumberStatus.Illegible
@@ -674,7 +505,7 @@ public class OcrServiceBase
             "\n",
             new AccountNumberRow
             {
-                Data = new Models.AccountNumber()
+                Data = new Business.Models.AccountNumber()
                 {
                     Number = "?00?000?1",
                     Status = AccountNumberStatus.Illegible
